# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "epubapi",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "crate::service::book::route"
        ],
        "summary": "閲覧可能なbook一覧を取得する",
        "description": "page: ページ番号\nkeyword: タイトル・著者名での検索キーワード",
        "operationId": "get_books",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            },
            "style": "form"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "style": "form"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "owner_id",
                      "name",
                      "creator",
                      "publisher",
                      "date",
                      "cover_image",
                      "visibility",
                      "direction",
                      "created_at",
                      "tags",
                      "epub_url"
                    ],
                    "properties": {
                      "cover_image": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date"
                      },
                      "creator": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string"
                      },
                      "direction": {
                        "type": "string",
                        "enum": [
                          "ltr",
                          "rtl"
                        ]
                      },
                      "epub_url": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "owner_id": {
                        "type": "string"
                      },
                      "publisher": {
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "visibility": {
                        "type": "string",
                        "enum": [
                          "public",
                          "private"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "unauthorized": "missing user id"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}": {
      "delete": {
        "tags": [
          "crate::service::book::route"
        ],
        "summary": "bookを削除する",
        "operationId": "delete_book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "unauthorized": "missing user id"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "crate::service::book::route"
        ],
        "summary": "bookを更新する",
        "operationId": "update_book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "visibility"
                ],
                "properties": {
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "unauthorized": "missing user id"
                }
              }
            }
          }
        }
      }
    },
...
    "/books/{book_id}/tags/{tag_name}": {
      "delete": {
        "tags": [
          "crate::service::book::route"
        ],
        "summary": "bookからtagを削除する",
        "operationId": "delete_tag_from_book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "unauthorized": "missing user id"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}